requirements.txt
huggingface_hub
pydantic
pyyaml
logzero
aiohttp